Remove Unused Code, Format
