--- compiler/coenobita-check/src/shared.rs
@@ -1,3 +1,0 @@
-use rustc_span::ErrorGuaranteed;
-
-pub type Result<T = ()> = std::result::Result<T, ErrorGuaranteed>;
