+++ compiler/coenobita-middle/src/lib.rs
@@ -1,13 +1,2 @@
-#![feature(rustc_private)]
-extern crate rustc_span;
-
-pub mod flow;
-pub mod origin;
-pub mod property;
-pub mod provenance;
-
 pub mod set;
-
 pub mod ty;
-
-pub mod map;
