--- compiler/coenobita-middle/src/map.rs
@@ -1,19 +1,0 @@
-use std::{collections::HashMap, hash::Hash};
-
-pub struct Map<F, T> {
-    map: HashMap<F, T>,
-}
-
-impl<F: Eq + Hash, T> Map<F, T> {
-    pub fn new() -> Self {
-        Map { map: HashMap::new() }
-    }
-
-    pub fn get(&self, key: &F) -> Option<&T> {
-        self.map.get(key)
-    }
-
-    pub fn insert(&mut self, key: F, value: T) {
-        self.map.insert(key, value);
-    }
-}
